site: split tfa config into its own page
