+++ core/Language/Language/Site/wizard.json
@@ -17,11 +17,11 @@
 		"Wizard_SchedulesCreate": "Schedule is to program the system to create Moves in the future, one or many, Daily, Monthly or Yearly. One can define for how long it will repeat, including non-stopping Schedules.",
 		"Wizard_SchedulesIndexEmpty": "Here will be listed the Schedules created inside the Account Areas. Enter in one Account to create Schedules, it's in the Area top menu.",
 		"Wizard_SchedulesIndexFilled": "These are the Schedules made in Account Areas. One can cancel the Schedule so it won't create Moves anymore.",
-		"Wizard_Settings": "Here you can change security and timezone information, the system theme you see and can enable/disable features. The enabled features make the system more complex, but give you more tools to control the money flow.",
 		"Wizard_SettingsMain": "Here you can change language, timezone and can enable/disable features. The enabled features make the system more complex, but give you more tools to control the money flow.",
 		"Wizard_SettingsPassword": "To change your password you need to remember the current one. In case you don't, click at Log off button at the top of the screen in the right at menu, go to login screen and click at 'Forgot my password'.",
 		"Wizard_SettingsEmail": "Here you can change your e-mail address at the system. Your e-mail is the only way we can contact you in case of problems or for notifications. Kept it up to date and be sure to have access to it.",
 		"Wizard_SettingsTheme": "The website and mobile app colors can be changed between three main colors and between light or dark theme. To make a better decision, at this screen you can see the controls used through the website. When you click at the buttons the colors change just so you can check and decide, to effective change one needs to click the Save button.",
+		"Wizard_SettingsTFA": "This page is about adding one more security layer to protect your login at the system. Click at 'Why to use this?' below the image to read more about this.",
 		"Wizard_MonthEmpty": "At this Area, you can see and create Moves. To add Moves, click at Move, at top menu.",
 		"Wizard_MonthFilled": "The Account Area start view list Moves by chosen month. It's accessible by top menu, at Extract. It's possible to access Year Summary at top menu too.",
 		"Wizard_Year": "Here you can see the year totals by month so you can have a general comparative year view.",
@@ -45,11 +45,11 @@
 		"Wizard_SchedulesCreate": "Esta tela serve para criar Movimentações futuras, seja apenas uma ocorrência ou várias repetidas. As frequências possíveis são Diária, Mensal e Anual. É possível definir por quantos dias/meses/anos será a repetição, incluindo Agendamentos sem um limite.",
 		"Wizard_SchedulesIndexEmpty": "Aqui são mostrados os Agendamentos feitos dentro da Área da Conta. Entre em uma Conta para criar um Agendamento, clicando em Agendar.",
 		"Wizard_SchedulesIndexFilled": "Estes são os Agendamentos feitos dentro das Áreas das Contas. É possível cancelar um Agendamento, para que não insira mais Movimentações.",
-		"Wizard_Settings": "Nesta tela é possível alterar informações de segurança e horário, o tema visto por você e habilitar e desabilitar funcionalidades. As funcionalidades habilitadas tornam o sistema mais complexo, mas fornecem mais ferramentas de controle do fluxo de dinheiro.",
 		"Wizard_SettingsMain": "Nesta tela é possível alterar idioma, fuso horário e habilitar e desabilitar funcionalidades. As funcionalidades habilitadas tornam o sistema mais complexo, mas fornecem mais ferramentas de controle do fluxo de dinheiro.",
 		"Wizard_SettingsPassword": "Para trocar sua senha por este formulário, é necessário lembrar a senha atual. Caso não se lembre, clique no botão de Sair no topo a direita no menu, vá a tela de login e clique em 'Esqueci minha senha'.",
 		"Wizard_SettingsEmail": "Aqui é possível trocar seu endereço de e-mail cadastrado no sistema. Seu e-mail é a única forma que podemos contatar você no caso de problemas ou para notificações. Mantenha ele atualizado e tenha certeza de ter acesso a ele.",
 		"Wizard_SettingsTheme": "As cores do site e do aplicativo podem ser alteradas entre três cores principais e se prefere um tema claro ou escuro. Para melhor decisão, nesta tela são mostradas os principais items usados pelo site. Ao clicar nos botões as cores mudam para visualização apenas, para manter a alteração é necessário clicar no botão de salvar.",
+		"Wizard_SettingsTFA": "Esta página serve para adicionar uma camada a mais de segurança ao seu cadastro no sistema. Clique em 'Por que usar?' abaixo da imagem para ler mais sobre.",
 		"Wizard_MonthEmpty": "Nesta área é possível ver e criar Movimentações. Para adicionar Movimentações, clique em Mover, no menu acima.",
 		"Wizard_MonthFilled": "A tela inicial da Área da Conta apresenta as Movimentações por mês escolhido. É acessível pelo menu Extrato. Também é possível ver o Resumo Anual, clicando no menu acima.",
 		"Wizard_Year": "Aqui são mostradas as somas do ano divididas por mês, assim é possível ter uma visão mais geral e comparativa do ano.",
