--- site/MVC/Views/Configs/Config.cshtml
@@ -1,84 +1,0 @@
-﻿@model ConfigsConfigModel
-@{
-	ViewBag.Title = Context.Translate("Settings");
-	ViewBag.WizardMessage = "Settings";
-	Layout = "~/Views/Shared/Layouts/_Form.cshtml";
-
-	var tabs = Enum.GetValues(typeof (ConfigsConfigModel.Form));
-
-	ViewBag.FormSizeSm = 8;
-	ViewBag.FormSizeLg = 6;
-}
-
-@section Styles {
-	<link rel="stylesheet" type="text/css" href="~/Assets/styles/qrcode.css?v=@Cfg.Version" />
-}
-
-@section Scripts {
-	<script type="text/javascript" src="~/Assets/scripts/tab-forms.js?v=@Cfg.Version"></script>
-	<script type="text/javascript" src="~/Assets/scripts/dynamic-theme-change.js?v=@Cfg.Version"></script>
-	<script type="text/javascript" src="~/Assets/libs/jquery/jquery.qrcode.min.js?v=@Cfg.Version"></script>
-	<script type="text/javascript" src="~/Assets/scripts/qrcode.js?v=@Cfg.Version"></script>
-	<script type="text/javascript">
-		window.currentBrightness = '@Model.Brightness';
-		window.currentColor = '@Model.Color';
-	</script>
-}
-
-<ul class="nav nav-tabs">
-	@foreach (ConfigsConfigModel.Form tab in tabs)
-	{
-		var name = tab.ToString();
-		var lower = name.ToLower();
-
-		var button = Context.Translate("Save", $"{name}Settings");
-
-		if (tab == ConfigsConfigModel.Form.TFA && Model.TFA.IsActive)
-		{
-			button = Context.Translate("TFARemove");
-		}
-
-		<li class="@lower-caller">
-			<a href="#@lower" data-toggle="tab">
-				<small class="tab-text" data-button="@button">
-					@Context.Translate($"{name}Settings")
-				</small>
-			</a>
-		</li>
-	}
-</ul>
-
-@Html.HiddenFor(m => m.ActiveForm)
-
-<div class="tab-content well-sm">
-	@foreach (var tab in tabs)
-	{
-		var name = tab.ToString();
-		var lower = tab.ToLower();
-
-		<div class="tab-pane" id="@lower" data-form="@Url.Action($"Config{name}")">
-			@await Html.PartialAsync($"Config{name}")
-		</div>
-	}
-</div>
-
-@section Footer {
-	@if (Model.TFA.IsActive)
-	{
-		var action = Model.TFA.TFAPassword
-			? "TFAPasswordDisable"
-			: "TFAPasswordEnable";
-
-		<small>
-			<a class="pull-left text-left"
-					href="@Url.Action(action, "Configs")">
-				@Html.Raw(Context.Translate(action).Replace("\n", "<br />"))
-			</a>
-		</small>
-	}
-
-	<button type="submit"
-			class="btn btn-success tab-submit">
-		&hellip;
-	</button>
-}
