--- site/MVC/Views/Configs/ConfigTFA.cshtml
@@ -1,39 +1,0 @@
-﻿@model ConfigsConfigModel
-
-@Html.HiddenFor(m => m.TFA.IsActive)
-
-@if (Model.TFA.IsActive)
-{
-	<div class="form-group">
-		@Html.LabelFor(m => m.TFA.TFA.Password, Context.Translate("CurrentPassword"), new { @class = "control-label" })
-		@Html.PasswordFor(m => m.TFA.TFA.Password, new { @class = "form-control" })
-	</div>
-}
-else
-{
-	<div class="text-center">
-		<span id="qrcode" data-url="@Model.TFA.UrlPath">
-		</span>
-
-		@Html.HiddenFor(m => m.TFA.TFA.Secret)
-
-		<br/>
-		<small>
-			<a href="#" role='button' data-toggle='modal' data-target='#tfa-modal'>
-				@Context.Translate("TFAWhy")
-			</a>
-		</small>
-	</div>
-
-	<div class="form-group">
-		@Html.LabelFor(m => m.TFA.TFA.Code, Context.Translate("Code"), new { @class = "control-label" })
-		@Html.TextBoxFor(m => m.TFA.TFA.Code, new { @class = "form-control" })
-	</div>
-
-	<div class="form-group">
-		@Html.LabelFor(m => m.TFA.TFA.Password, Context.Translate("CurrentPassword"), new { @class = "control-label" })
-		@Html.PasswordFor(m => m.TFA.TFA.Password, new { @class = "form-control" })
-	</div>
-}
-
-@await Html.PartialAsync("Modals/TFA")
