+++ site/Tests/Browser/configs.test.js
@@ -7,13 +7,6 @@ describe('Configs', () => {
 		user = await puppy.logon('configs@dontflymoney.com')
 	})
 
-	test('Visit', async () => {
-		await puppy.call('Configs/Config')
-
-		const header = await puppy.content('.panel .header')
-		await expect(header).toContain('Configurações')
-	})
-
 	test('Index', async () => {
 		await puppy.call('')
 		await page.click('#settings')
@@ -68,5 +61,19 @@ describe('Configs', () => {
 
 		const message = await puppy.content('.alert')
 		await expect(message).toContain('Configurações alteradas')
+	})
+
+	test('TFA', async () => {
+		await puppy.call('')
+		await page.click('#settings')
+		await page.click('#settings_tfa', { visible: true })
+
+		const header = await puppy.content('.panel .header')
+		await expect(header).toContain('Login mais seguro')
+
+		await puppy.submit(`/Configs/TFA`)
+
+		const message = await puppy.content('.alert')
+		await expect(message).toContain('Código inválido')
 	})
 })
